requirements.txt
Django~=5.0.7
djangorestframework~=3.15.2
django-environ~=0.11.2

django_cors_headers~=4.4.0
django-filter~=24.2

drf-spectacular[sidecar]~=0.27.2
django-debug-toolbar~=4.4.6

black~=24.4.2
ruff~=0.5.1
